Database migrations are now handled by Alembic instead of Flyway. Updated init_db utiliy and added __main__.py to make it callable as a Python module. Updated tests and moved app & db fixtures into conftest.py.

+++ backend_main/db/init_db.py
@@ -1,22 +1,16 @@
-"""
-Database initialization + migration module.
-"""
-import os, sys
-import subprocess
+import os
 
 import psycopg2
+import alembic.config
 
-if __name__ == "__main__":
-    sys.path.insert(0, os.path.dirname(os.path.dirname(sys.path[0])))
 
-from backend_main.config import get_config
-    
+class DBExistsException(Exception):
+    pass
 
 def connect(host, port, database, user, password):
-    connection = psycopg2.connect(host = host, port = port, database = database, \
-                            user = user, password = password)
+    connection = psycopg2.connect(host=host, port=port, database=database, \
+                            user=user, password=password)
     connection.set_session(autocommit=True)
-    
     return connection.cursor()
 
 
@@ -35,16 +29,24 @@ def create_user(cursor, user, password):
                             RAISE NOTICE 'Not creating {user} role, because it already exists.';
                         END
                     $$;""")
-    while cursor.connection.notices:
-        print(cursor.connection.notices.pop())
-    print("Finished creating the user.")
+    if len(cursor.connection.notices) > 0:
+        while cursor.connection.notices:
+            print(cursor.connection.notices.pop().rstrip())
+    else:
+        print("Finished creating the user.")
 
 
-def create_db(cursor, db_name, db_owner):
-    # Close connections and drop db if it exists
+def create_db(cursor, db_name, db_owner, force):
+    # Check if db exists
     cursor.execute(f"SELECT COUNT(*) as count FROM pg_database WHERE datname='{db_name}'")
     db_exists = cursor.fetchone()[0]
+
     if db_exists:
+        # Stop if db exists and --force flag is not provided
+        if not force:
+            raise DBExistsException(f"Database already exists, exiting without recreating and applying migrations.")
+        
+        # Close connections and drop existing database if --force flag is provided
         cursor.execute(f"""
                         SELECT pg_terminate_backend(pg_stat_activity.pid)
                         FROM pg_stat_activity
@@ -52,74 +54,53 @@ def create_db(cursor, db_name, db_owner):
                         AND pid <> pg_backend_pid();
         """)
         cursor.execute(f"DROP DATABASE {db_name};")
+        print(f"Deleted existing database '{db_name}'.")
     
     cursor.execute(f"CREATE DATABASE {db_name} ENCODING 'UTF-8' OWNER {db_owner} TEMPLATE template0;")
     print("Finished creating the database.")
 
 
-def create_schema(host, port, database, user, password, schema):
-    cursor = None
-    try:
-        cursor = connect(host, port, database, user, password)
-        cursor.execute(f"CREATE SCHEMA IF NOT EXISTS {schema} AUTHORIZATION {user};")
-        print("Finished creating the schema.")
-    finally:
-        disconnect(cursor)
-    
+def revision(message = ""):
+    # Set current working directory
+    cwd = os.getcwd()
+    alembic_dir = os.path.dirname(__file__)
+    os.chdir(alembic_dir)
 
-def create_flyway_conf(db_config, folder):    
-    conf_file = folder + "/flyway.conf"
+    # Run revision command
+    alembic_args = ["revision", "--autogenerate", f'-m "{message}"']
+    alembic.config.main(argv=alembic_args)
 
-    host = db_config["db_host"]
-    port = db_config["db_port"]
-    database = db_config["db_database"]
-    user = db_config["db_username"]
-    password = db_config["db_password"]
-    schema = db_config["db_schema"]
-
-    if not os.path.exists(folder):
-        os.mkdir(folder)
-    
-    if os.path.exists(conf_file):
-        os.remove(conf_file)
-        
-    with open(conf_file, "w") as write_stream:
-        write_stream.write("\n".join([
-           f"flyway.url=jdbc:postgresql://{host}:{port}/{database}",
-           f"flyway.user={user}",
-           f"flyway.password={password}",
-           f"flyway.schemas={schema}",
-           f"flyway.locations=filesystem:./"
-        ]))
-    
-    print("Finished creating flyway.conf file.")
+    # Restore current working directory
+    os.chdir(cwd)
+    print("Finished database revision.")
 
 
-def migrate_db(folder):
-    p = subprocess.Popen(["flyway", "migrate"], shell = True, cwd = folder)
-    p.wait()
+def migrate(config_file = None):
+    # Set current working directory
+    cwd = os.getcwd()
+    alembic_dir = os.path.dirname(__file__)
+    os.chdir(alembic_dir)
+
+    # Run revision command
+    alembic_args = ["-x", f'app_config_path="{config_file}"'] if config_file else []
+    alembic_args.extend(["upgrade", "head"])
+    alembic.config.main(argv=alembic_args)
+
+    # Restore current working directory
+    os.chdir(cwd)
+    print("Finished migrating the database.")
 
 
-def init_db():
-    cursor = None
-    
-    try:
-        db_config  = get_config()["db"]
-        cursor = connect(host = db_config["db_host"], port = db_config["db_port"], database = db_config["db_init_database"],
-                            user = db_config["db_init_username"], password = db_config["db_init_password"])
-        if db_config["create_user_required"]:
-            create_user(cursor, db_config["db_username"], db_config["db_password"])
-        create_db(cursor, db_config["db_database"], db_config["db_username"])
-        create_schema(host = db_config["db_host"], port = db_config["db_port"], database = db_config["db_database"],
-                        user = db_config["db_username"], password = db_config["db_password"], 
-                        schema = db_config["db_schema"])
-    finally:
-        disconnect(cursor)
-    
-    migrations_folder = os.path.dirname(os.path.abspath(__file__)) + "/migrations"
-    create_flyway_conf(db_config, migrations_folder)
-    migrate_db(migrations_folder)
+# def migrate():
+#     # Set current working directory
+#     cwd = os.getcwd()
+#     alembic_dir = os.path.dirname(__file__)
+#     os.chdir(alembic_dir)
 
+#     # Run revision command
+#     alembic_args = ["upgrade", "head"]
+#     alembic.config.main(argv=alembic_args)
 
-if __name__ == "__main__":
-    init_db()
+#     # Restore current working directory
+#     os.chdir(cwd)
+#     print("Finished migrating the database.")
